various housekeeping
